move `toStaticUrl()` helper to @depmap/globals package
